+++ Game/MapManager.cs
@@ -44,9 +44,9 @@ namespace UAlbion.Game
                 var map = new Map2D(_pendingMapChange.Value);
                 Raise(new SetSceneEvent((int)SceneId.World2D)); // Set the scene first to ensure scene-local components from other scenes are disabled.
                 exchange.Attach(map);
-                Raise(new SetInputModeEvent((int)InputMode.World2D));
+                Raise(new SetInputModeEvent(InputMode.World2D));
                 Raise(new CameraJumpEvent((int)map.LogicalSize.X / 2, (int)map.LogicalSize.Y / 2));
-                Raise(new LogEvent((int)LogEvent.Level.Info, $"Loaded map {(int)_pendingMapChange}: {_pendingMapChange}"));
+                Raise(new LogEvent(LogEvent.Level.Info, $"Loaded map {(int)_pendingMapChange}: {_pendingMapChange}"));
             }
 
             var mapData3D = assets.LoadMap3D(_pendingMapChange.Value);
@@ -56,9 +56,9 @@ namespace UAlbion.Game
                 var map = new Map3D(_pendingMapChange.Value);
                 Raise(new SetSceneEvent((int)SceneId.World3D)); // Set the scene first to ensure scene-local components from other scenes are disabled.
                 exchange.Attach(map);
-                Raise(new SetInputModeEvent((int)InputMode.MouseLook));
+                Raise(new SetInputModeEvent(InputMode.MouseLook));
                 Raise(new CameraJumpEvent((int)map.LogicalSize.X / 2, (int)map.LogicalSize.Y / 2));
-                Raise(new LogEvent((int)LogEvent.Level.Info, $"Loaded map {(int)_pendingMapChange}: {_pendingMapChange}"));
+                Raise(new LogEvent(LogEvent.Level.Info, $"Loaded map {(int)_pendingMapChange}: {_pendingMapChange}"));
             }
 
             _pendingMapChange = null;
